Lock debugging: Fix ConcurrentModificationException
unsynchronized access to the readers_ property causes a said exception. Since
the super class synchronizes on this, let's do the same here. It should be
sufficient to synchronize only when getting the iterator and looping through it,
but that would change more lines (indentation), therefore I chose to
synchronize the whole method.